+++ src/main/scala/ai/starlake/schema/generator/Yml2DDLJob.scala
@@ -119,7 +119,8 @@ class Yml2DDLJob(config: Yml2DDLConfig, schemaHandler: SchemaHandler)(implicit
                 None
               ),
               connectionOptions,
-              skipRemarks = true
+              skipRemarks = true,
+              keepOriginalName = true
             )
           case None => Map.empty[String, (TableRemarks, Columns, PrimaryKeys)]
         }