+++ CHANGELOG.md
@@ -15,6 +15,7 @@ __Improvement__:
 - added `auditConnectionRef` to jdbc extract schemas to be on pair with connectionRef behavior
 - warn when starlake version is not installed yet and user want to use it
 - added support for mysql extraction
+- add the ability to rename column during schema and data extraction
 
 # 1.0.0:
 - **BREAKING CHANGE**